% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightHorizons.R
\name{weightHorizons}
\alias{weightHorizons}
\title{Weight SSURGO Horizon data into mapunits}
\usage{
weightHorizons(x, horizonVar)
}
\arguments{
\item{x}{a SSURGO list which is returned by running the \code{get_ssurgo} function from the fedData package}

\item{horizonVar}{A variable from the 'chorizon' SSURGO table. A list can be found here: https://www.nrcs.usda.gov/wps/PA_NRCSConsumption/download?cid=stelprdb1241114&ext=pdf Make sure to substitute '.' (periods) for '_' (underscores) in the column names.}
}
\value{
A simple features multi-polygon layer which contains a columns called 'weightedMeanVar' which will show the weighted variable you specified for each mapunit.
}
\description{
Create a a flat file that takes values from the SSURGO chorizon table and
weighs them by what percent of a component each horizon composes and also what percent of a mapunit each component makes up.
}
\examples{
library(USAboundaries)
library(sf)
library(tidyverse)
library(FedData)
ok <- us_states(resolution = "low")\%>\%   # Download all state boundaries
filter(name == "Oklahoma")\%>\%   # Filter to Oklahoma
  st_transform(crs = 32614) # Project to UTM 14N so distance units are in meters

okgrid <- st_make_grid(ok,cellsize = 20000,    # 20,000 m cell size
                       square = FALSE)\%>\% # Returns sfc_POLYGON
  st_as_sf()
okgrid$ID <- paste0("pt_",rownames(okgrid))
colnames(okgrid) <- c("geometry","ID")
 st_geometry(okgrid) <- "geometry"
sa <- okgrid[365,]\%>\%
   mutate(ID = "Ada")
ada <- as_Spatial(sa)
ssurgo <- get_ssurgo(ada, label = "Ada")
weighted <- weightHorizons(ssurgo, "ksat.r")

plot(weighted["weightedMeanVar"], main = "Weighted ksat")
}
